pyproject.toml
[tool.poetry]
name = "saude"
version = "0.1.0"
description = ""
authors = ["matheusgb <matheusgbr98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"